add restriction for student

+++ app/Http/Controllers/listingStudentController.php
@@ -10,6 +10,8 @@ use App\Models\Section;
 use App\Models\Student;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+
 
 class listingStudentController extends Controller
 {
@@ -23,30 +25,41 @@ class listingStudentController extends Controller
 
     }
     public function showUserInfo($id){
-        // dd($id);
 
-        $studentid = Student::where('user_id','=',$id)->get();  
+        $response = Gate::inspect('dashboard-formation-suivi');
         
-        $sections = Section::where('id','=',$studentid->first()->section_id)->get();
-        
-        // dd($sections);  
-        $FollowUps = FollowUp::where('student_id','=',$studentid->first()->id)->get();
-        // dd($FollowUps);
-        $User = User::where('id','=',$id)->get();
-        // $studentclass = Student::select('section_id')->where('id','=',$id)->get();    
+        if ($response->allowed()) {
+            $studentid = Student::where('user_id','=',$id)->get();  
+            
+            $sections = Section::where('id','=',$studentid->first()->section_id)->get();
+            
+            $FollowUps = FollowUp::where('student_id','=',$studentid->first()->id)->get();
+            
+            $User = User::where('id','=',$id)->get();
+            // $studentclass = Student::select('section_id')->where('id','=',$id)->get();    
 
-        // $sections = Section::select('class_name')->where("id",'=', $studentclass->first()->section_id)->get();
-        // dd($sections);
-        // dd($student,$User,$FollowUps);
-        return view('adviser.listingOneStudent', compact('FollowUps','User','studentid','sections'));
+            // $sections = Section::select('class_name')->where("id",'=', $studentclass->first()->section_id)->get();
+            
+            return view('adviser.listingOneStudent', compact('FollowUps','User','studentid','sections'));
+        } else {
+            echo $response->message();
+        }
 
     }
 
      // Function to delete one alternant
      public function deleteOneAlternant($id) {
-        $post = Student::findOrFail($id);
-        $post->delete();
-        return redirect()->back()->with('success', 'Alternant supprimé avec succès.');
+
+        $response = Gate::inspect('listingAlternant.delete');
+
+        if ($response->allowed()) {
+
+            $post = Student::findOrFail($id);
+            $post->delete();
+            return redirect()->back()->with('success', 'Alternant supprimé avec succès.');
+        } else {
+            echo $response->message();
+        }
     }
 
 }
